removed additional items commented line 456 static void needed / in lines 10 & 11 since working on a Linux system.
